📦 Opcodes `0x08-0x0F` Added
- New Opcodes

+++ include/CPU.h
@@ -36,7 +36,7 @@ private:
     void executeInstructions(int cycles);
 
     // Flag Methods
-    void checkCarry(uint16_t prev, uint16_t after);
+    void checkCarry(uint8_t prev, uint16_t after);
     void checkHalfCarry(uint8_t prev, uint8_t after);
 
     // Initializers
