+++ include/Memory.h
@@ -7,9 +7,9 @@
 
 class Memory {
 private:
-    uint16_t memory[0xFFFF];
+    uint8_t memory[0xFFFF];
     uint8_t cartType;
-    std::stringstream disassembledROM;
+    std::stringstream disassembledROM;    // TODO: Byte Array instead!
     void dumpROM(std::string ROMPath, std::ostream &out);
 public:
     Memory(std::string ROMPath);
