+++ src/CPU.cc
@@ -31,12 +31,12 @@ CPU::~CPU() {
 
 
 /**
- * Sets Half-Carry flag if a half carry occured
+ * Sets Half-Carry flag if a half carry occurred for 8bit Value
  * @param prev The previous value
  * @param after Result of the previous value
  * @return Result of the Set (if needed)
  */
-void CPU::checkHalfCarry(uint8_t prev, uint8_t after) {
+void CPU::checkHalfCarry(uint8_t prev, uint8_t after) { // Overflow from 3rd Bit
   this->reg.F &= 0xFD;                    // Unset H Flag
   if ((prev & 0x0F) != (after & 0x0F))    // LSB Change
     if ((prev & 0xF0) != (after & 0xF0))  // MSB Change
@@ -44,12 +44,12 @@ void CPU::checkHalfCarry(uint8_t prev, uint8_t after) {
 }
 
 /**
- * Sets Carry flag if a carry occured
+ * Sets Carry flag if a carry occurred for 8bit Value
  * @param prev The previous value
  * @param after Result of the previous value
  * @return Result of the Set (if needed)
  */
-void CPU::checkCarry(uint16_t prev, uint16_t after) {
+void CPU::checkCarry(uint8_t prev, uint16_t after) { // Overflow from 7th Bit
   this->reg.F &= 0xFE;                        // Unset C Flag
   if ((prev & 0xFF) != (after & 0xFF))        // LSB Change
     if ((prev & 0xFF00) != (after & 0xFF00))  // MSB Change
